% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_oubtbreak.R
\docType{function}
\name{create_outbreak}
\alias{create_outbreak}
\title{Create an outbreak}
\usage{
create_outbreak(size = 500, startDate = "2015-01-01",
  endDate = "2015-12-31", dateOfInfection = "2015-05-06",
  centerNames = c("Oranienburger Str", "Buchholzerstr",
  "Platz der Luftbruecke", "Bizetstr"), meanStayAtCenter = 50,
  disease = "varicella", diseaseIncubationPeriod = 14)
}
\arguments{
\item{size}{The total number of the population. A number.}

\item{startDate}{The start of the period in which the outbreak takes place. A chracter vector of a date e.g. "2015-01-01"}

\item{endDate}{The end of the period in which the outbreak takes place. A chracter vector of a date e.g. "2015-12-31"}

\item{dateOfInfection}{The date the outbreak begins. . A chracter vector of a date e.g. "2015-06-01"}

\item{centerNames}{The names of the centers at which the outbreak takes place. A character vector}

\item{meanStayAtCenter}{The mean time a person spends at a center}

\item{disease}{The name of the disease. A character vector.}

\item{diseaseIncubationPeriod}{The incubation period of a disease}
}
\description{
Create an outbreak
}
\section{Background}{

Infect the world - or at least create a dataset of an varicella outbreak in centers for foreigners

This dataset simulates an outbreak of varicella in german centers for foreigners. It is loosely based on
the situation in 2015, when the numbers of foreigners seeking asylum exeded the available places
in the center for foreigners. Varicella was the most frequent disease in these centers at that time.
comparable with kindergartens and other shelters.

The data is useful to compute incidence rates.
}

\examples{

create_outbreak()

}
\references{
Bericht über die dem Robert Koch-Institut übermittelte meldepflichtige Infektionskrankheiten bei Asylsuchenden in Deutschland
https://www.rki.de/DE/Content/Gesundheitsmonitoring/Gesundheitsberichterstattung/GesundAZ/Content/A/Asylsuchende/Inhalt/meldepflichtige_Infektionskrankheiten_bei_Asylsuchenden.pdf?__blob=publicationFile
}
\author{
Jakob Schumacher (\email{jakob.schumacher@web.de})
}
